requirements.txt
click==8.1.3
discord==2.2.3
discord.py==2.2.3
InquirerPy==0.3.4
Pillow==9.5.0
Pillow==9.5.0
psutil==5.9.5
pycryptodome==3.17
pyfiglet==0.8.post1
pyobf2==1.2.0
pywin32==306
Requests==2.30.0
rich==13.3.5
WMI==1.5.1
urllib3==2.0.2
attrs==23.1.0